app/the_dipper/requirements.txt
glitch-this
Pillow~=8.3.2
ffmpeg-python
opencv-python~=4.5.4.60
imutils